fix: deduplicate the definition of Tekton pipelines
The whole `pipelineSpec` subtree (as provided by the Konflux generator)
is identical for the `pull-request` and `push` pipelines.  It does not
make any sense to change everything in two separate files that are
*almost* identical with each other.  This commit merges both of them
in a single file so that we can use a YAML anchor to eliminate the
excessive duplication.

Related: https://issues.redhat.com/browse/AEGIS-118