fix: rename `rewrite-{description,statement}`
... to `suggest-{description,statement}`, respectively.

The previous names of the `rewrite-{description,statement}` features
were misleading because in vast majority of cases, the description and
statement fields in OSIM are empty when these Aegis features are used.

This change makes the interface more intuitive.  Doing the change now,
when the corresponding API endpoints are not yet used, is much easier
because we do not need to care about keeping backward compatibility
with existing users of the API.

Related: https://issues.redhat.com/browse/AEGIS-212