evals: tweak the versioning info detection rubric
Before this change, the `suggest-description-for-CVE-2025-54770`
evaluation case occasionally failed with:
```
The 'suggested_description' contains 'version 2', which is versioning information. The rubric states that neither suggested_title nor suggested_description should contain versioning info.
```
... because of the expanded `GRUB2` acronym in `suggested_description`:
```
A flaw was found in GRUB2 (Grand Unified Bootloader version 2).  This vulnerability allows a Denial of Service (DoS) via a use-after-free issue when the `net_set_vlan` command is invoked after the network module is unloaded, leading to system instability.
```