evals: tweak the versioning info detection rubric
Before this change, the `suggest-description-for-CVE-2025-54770`
evaluation case was still occasionally failing with:
```
The suggested_title contains "GRUB2", which includes versioning information ("2"), and this is not part of an acronym explanation within the title. The rubric states that versioning info is not allowed except in acronyms explanations.
```

Fixes: commit 6cd03469ecbfee6613f3f5ebd20b38bdb5fddb83
Related: https://issues.redhat.com/browse/AEGIS-264